requirements.txt
amqp==2.2.2
anyjson==0.3.3
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
billiard==3.5.0.3
celery==4.2.0rc1
cffi==1.11.5
click==6.7
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
decorator==4.2.1
Django==1.9.13
django-celery==3.2.1
django-dynamic-scraper==0.13.0
enum34==1.1.6
freeze==1.0.10
freeze-requirements==0.5.3
future==0.15.2
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipaddress==1.0.19
jsonpath-rw==1.4.0
kombu==4.1.0
lxml==4.2.1
parsel==1.4.0
pillow==3.4.2
ply==3.11
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.5.0
pytz==2018.3
queuelib==1.5.0
Scrapy==1.4.0
scrapy-djangoitem==1.1.1
scrapy-splash==0.7.2
scrapyd==1.2.0
service-identity==17.0.0
sh==1.12.14
six==1.11.0
Twisted==17.9.0
vine==1.1.4
w3lib==1.19.0
zope.interface==4.4.3
